requirements.txt
playwright==1.42.0
python-telegram-bot==21.0.1
python-dotenv==1.0.1
aiosqlite==0.20.0
beautifulsoup4==4.12.3
httpx==0.27.0
playwright-stealth==1.0.6
loguru==0.7.2
